added gp7/8 import (#59)

+++ src/main/java/log/charter/data/song/position/fractional/IConstantFractionalPosition.java
@@ -19,7 +19,7 @@ public interface IConstantFractionalPosition extends IVirtualConstantPosition {
 		return this;
 	}
 
-	default int position(final ImmutableBeatsMap beats) {
+	default double position(final ImmutableBeatsMap beats) {
 		return position().getPosition(beats);
 	}
